ref: Extracts code from Endpoint to a separate class.
Extracts the code for handling sending and receiving messages for an
Endpoint into EndpointMessageTransport in preparation for the
introduction of a new underlying transport (web sockets).
